Refactor AsyncEditorFormComponent into separate stateless and stateful components.
